# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Coaching planner backend Documentation",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "rubenvdkamp08@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Coaching planner API Server"
        },
        {
            "url": "http://my-default-host.com",
            "description": "Coaching planner API Server"
        },
        {
            "url": "http://my-default-host.com",
            "description": "Coaching planner API Server"
        },
        {
            "url": "http://coachingplanner.test"
        }
    ],
    "paths": {
        "/agenda": {
            "post": {
                "tags": [
                    "Agenda"
                ],
                "summary": "Get agenda of coach",
                "description": "Returns agenda of coaches to build agenda frontend",
                "operationId": "get_agenda",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    [
                        "coachingplanner_auth_key"
                    ]
                ]
            }
        },
        "/agenda/update": {
            "post": {
                "tags": [
                    "Agenda"
                ],
                "summary": "Update agenda of coach",
                "description": "Update, create, delete agenda moment of coach",
                "operationId": "update_agenda",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    [
                        "coachingplanner_auth_key"
                    ]
                ]
            }
        },
        "/agenda/moments/{user}": {
            "get": {
                "tags": [
                    "Agenda"
                ],
                "summary": "Get agenda moments of coach per half an hour",
                "description": "Returns agenda moments per day and per half an hour of coaches for students of coach",
                "operationId": "get_agenda_moments_coach",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    [
                        "coachingplanner_auth_key"
                    ]
                ]
            }
        },
        "/agenda/moment/{user}": {
            "post": {
                "tags": [
                    "Agenda"
                ],
                "summary": "Assign coaching moment to student",
                "description": "Assigning coaching moment to student of the coach",
                "operationId": "assign_agenda_moment_to_student",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AgendaMomentRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    [
                        "coachingplanner_auth_key"
                    ]
                ]
            }
        },
        "/agenda/next": {
            "get": {
                "tags": [
                    "Agenda"
                ],
                "summary": "Get next coaching moment",
                "description": "Returns details of the next coaching moment for the coach",
                "operationId": "get_next_agenda_moment",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    [
                        "coachingplanner_auth_key"
                    ]
                ]
            }
        },
        "agenda/students/notify": {
            "get": {
                "tags": [
                    "Agenda"
                ],
                "summary": "Notify students to assign to coaching moment",
                "description": "Sends mail to all the students of the coach to assign for a coaching moment",
                "operationId": "notify_students",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    [
                        "coachingplanner_auth_key"
                    ]
                ]
            }
        },
        "agenda/overview/week": {
            "get": {
                "tags": [
                    "Agenda"
                ],
                "summary": "Get week overview of the coaching moments",
                "description": "Returns number per day of the week with the amount of coaching moments",
                "operationId": "agenda_overview_week",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    [
                        "coachingplanner_auth_key"
                    ]
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "AgendaMomentRequest": {
                "title": "Assign agenda moment request",
                "description": "Assign agenda moment to student of coach request body data",
                "required": [
                    "guid",
                    "start_time",
                    "end_time"
                ],
                "properties": {
                    "guid": {
                        "title": "guid",
                        "description": "Guid of the student that want to assign",
                        "type": "string",
                        "example": "9c058e90-7ef1-11ea-a2bd-6db2999910ca"
                    },
                    "start_time": {
                        "title": "start_time",
                        "description": "Start time of the assigned moment",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-03-10 15:00:00"
                    },
                    "end_time": {
                        "title": "end_time",
                        "description": "End time of the assigned moment",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-03-10 15:30:00"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "coachingplanner_auth_key": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}